Move RTK and Formik error handling for Metabot into utils.ts
