+++ enterprise/frontend/src/metabase-enterprise/metabot/components/MetabotAdmin/MetabotTrialPage.tsx
@@ -1,7 +1,5 @@
 import { useDisclosure } from "@mantine/hooks";
-import type { FetchBaseQueryError } from "@reduxjs/toolkit/query";
 import { useCallback } from "react";
-import { P, isMatching } from "ts-pattern";
 import { t } from "ttag";
 import * as Yup from "yup";
 
@@ -23,48 +21,7 @@ import { Divider, Flex, Group, List, Stack, Text } from "metabase/ui";
 import { usePurchaseCloudAddOnMutation } from "metabase-enterprise/api";
 
 import { MetabotSettingUpModal } from "./MetabotSettingUpModal";
-
-// https://redux-toolkit.js.org/rtk-query/usage/error-handling
-// https://redux-toolkit.js.org/rtk-query/usage-with-typescript#type-safe-error-handling
-const isFetchBaseQueryError = (error: unknown): error is FetchBaseQueryError =>
-  isMatching({ status: P.any, data: P.any }, error);
-
-type IFieldError =
-  | string
-  | {
-      message: string;
-    }
-  | {
-      errors: { [key: string]: any };
-    };
-
-const isFieldError = (error: unknown): error is IFieldError =>
-  isMatching(
-    P.union(
-      P.string,
-      { message: P.string },
-      { errors: P.record(P.string, P.any) },
-    ),
-    error,
-  );
-
-export const handleFieldError = (error: unknown) => {
-  if (!isFieldError(error)) {
-    return;
-  }
-
-  if (typeof error === "string") {
-    throw { data: { errors: { terms_of_service: error } } };
-  }
-
-  if ("message" in error) {
-    throw { data: { errors: { terms_of_service: error.message } } };
-  }
-
-  if ("errors" in error) {
-    throw { data: error };
-  }
-};
+import { handleFieldError, isFetchBaseQueryError } from "./utils";
 
 interface MetabotTrialFormFields {
   terms_of_service: boolean;
